eatvul-webapp/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
torch==2.1.0
transformers==4.35.2
scikit-learn==1.3.2
google-generativeai==0.8.3
python-multipart==0.0.6
pydantic==2.5.0
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
python-dotenv==1.0.0 